\name{Menu.addcenter}
\alias{Menu.addcenter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Adding center points to a 2-level design}
\description{
  This help file describes adding center points to a 2-level design 
  from package \pkg{FrF2}.
}
%- maybe also 'usage' for other objects documented here.
\details{
  A name for the augmented design must be given.

  The number of center points is needed.
  
  Center points are distributed systematically over the design - they 
  are not randomized. The bottom text box requests the number of positions over which 
  the center points are to be distributed.
  
}

\seealso{See also function \code{\link[FrF2]{add.center}} for the 
function that does the work}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
