\name{Menu.BsMDPlot}
\alias{Menu.BsMDPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for 2-level Bayesian Screening analysis}
\description{Using the menu for 2-level Bayesian Screening analysis}
%- maybe also 'usage' for other objects documented here.
\section{Using the menu for 2-level Bayesian Screening analysis}{
  Select the response that is to be analysed.\cr
  Select the factors to include in the screening - usually use all factors, 
  but exclude dummy factors for pb type designs.\cr
  Select the degree of interactions to be included in the model.\cr
  Indicate, whether only a plot is requested or also a printout of model 
  probabilities. 
  The default is to also print the model probabilities.
}
\seealso{See also \code{\link[FrF2]{BsProb.design}} and \code{\link[BsMD]{plot.BsProb}} 
for the underlying functions.}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
