\name{Menu.colremove}
\alias{Menu.colremove}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for deleting columns}
\description{Help on using the menu for deleting columns}
%- maybe also 'usage' for other objects documented here.
\section{Help on using the menu for deleting columns}{
  This menu allows to permanently remove certain types of columns 
  (NOT: block factor column or 
  columns with experimental factors) from the design. 
  
  If you want to keep a response column in the data frame 
  but not treat it as a response variable any more,  
  use the select / deselect responses menu instead.
}

\seealso{See also \code{\link[DoE.base]{class-design}} for the underlying function}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
