\name{Menu.display}
\alias{Menu.display}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using display design menu}
\description{
  This help file describes displaying the active design in actual or standard 
  run order.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes displaying the active design in actual or standard 
  run order.

  Displaying in actual run order is the default. It is useful for feasibility 
  considerations, for   identifying unwanted systematic patterns, 
  or for identifying run order effects in the responses.
  
  Displaying in standard run order is useful, because it helps to identify structure 
  for some types of design (mainly for factorial designs, not so exciting e.g. for 
  latin hypercube designs). 
  }
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
