\name{Menu.EffectsPlots}
\alias{Menu.EffectsPlots}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for (half) normal effects plots}
\description{Help on using the menu for (half) normal effects plots}
%- maybe also 'usage' for other objects documented here.
\section{Help on using the menu for (half) normal effects plots}{
  Select the response that is to be analyzed.\cr
  Decide with the check buttons, \cr
  whether a half or full normal plot is requested, \cr
  whether effects are labelled with their names or with codes (usually \code{A}, \code{B} etc.)\cr
  and whether only significant effects are to be labelled\cr
  Change the value for alpha (level of significance) to your liking.
}

\seealso{See also \code{\link[FrF2]{DanielPlot}} for the underlying function}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
