\name{Menu.import}
\alias{Menu.import}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the import menu}
\description{
  This help file describes importing a design that is already in the R 
  workspace but is not the currently active data set.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes importing a design that is already in the R 
  workspace but is not the currently active data set.
  The dialog is only available 
  if at least one (object of class) design exists in the R workspace.
  
  Select the design from the drop-down list of available designs, and press OK.

  For importing an unloaded R work space first, use the 
  \emph{Load R workspace} menu item. 
  

}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
