\name{Menu.loadcatlg}
\alias{Menu.loadcatlg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the menu for loading catalogues}
\description{
  This help file describes loading a catalogue from a loaded catalogues package. 
  Currently, the only such package is FrF2.catlg128.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes loading a catalogue 
  package is \pkg{FrF2.catlg128}, if that package is installed.

  The menu is only available if package \pkg{FrF2.catlg128} is installed.
  
  Select the design from the drop-down list of available designs, and press OK.

  If R package \pkg{FrF2.catlg128} is not installed, it has to be installed
  outside of the R-commander. 
  
}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
