\name{Menu.mExport}
\alias{Menu.mExport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the export menu}
\description{
  This help file describes the menu items of the export menu.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes the menu items of the export menu.
  
  It is possible to

  change the working directory
    
  save the commands from the script window

  save the output from the output window

  save the R workspace
  
  export the active experimental design to \code{rda} format and - if desired - 
  to a csv file and/or an html file.  
}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
