\name{Menu.mImport}
\alias{Menu.mImport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the import menu}
\description{
  This help file describes the menu items of the import menu.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes the menu items of the import menu.
  
  It is possible to

  change the working directory
    
  Load a stored R workspace
  
  Load a design from the R workspace
  
  and to add response data from a csv file to an experimental design that 
  is already in the workspace or stored as an rda file.  
  
  Importing data - of course also experimental data - from other file formats is possible 
  from R commanders data management menu.
}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
