\name{Menu.plot}
\alias{Menu.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for plotting selected factors}
\description{Help on using the menu for plotting selected factors}
%- maybe also 'usage' for other objects documented here.
\section{Help on using the menu for plotting selected factors}{
  This menu allows to select experimental factors for plotting. 
  
  The menu is particularly useful for viewing the structure of 
  general orthogonal arrays. 
  In mosaic plots, up to three or four factors can be easily viewed together.
}

\seealso{See also the \code{\link[DoE.base:class-design-methods]{plot}} method for 
class \code{design} for the underlying function}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
