\name{Menu.QuantDesignAnalyze}
\alias{Menu.QuantDesignAnalyze}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Analyze designs for quantitative factors}
\description{This is a brief explanation on the menu for analysis of designs for quantitative factors.}
%- maybe also 'usage' for other objects documented here.

\section{Available analysis options}{
??? must be filled with content
}


\references{ 
    Box G. E. P, Hunter, W. C. and Hunter, J. S. (2005) 
    \emph{Statistics for Experimenters, 2nd edition}.
    New York: Wiley.
 }
\author{ Ulrike Groemping }
\seealso{ See Also ... 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
