\name{Menu.responses}
\alias{Menu.responses}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for response selection}
\description{Help on using the menu for response selection}
%- maybe also 'usage' for other objects documented here.
\section{Help on using the menu for response selection}{
  This menu allows to select numerical variables to become responses 
  or to deselect current responses so that they remain on the data frame 
  but loose their response status. 
  
  If you want to permanently remove a response altogether, 
  use the remove columns menu instead.
}

\seealso{See also \code{\link[DoE.base]{class-design}} for the underlying function}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
