\name{Menu.rsmmodel}
\alias{Menu.rsmmodel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the response surface model menu}
\description{
  This help file describes usage of the response surface model menu.
}
\usage{
  Menu.rsmmodel() 
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes usage of the response surface model menu. 
  The menu is only available 
  if the active data set is a design with at least one response.
  
  Select the intended response in the menu (it can also be changed afterwards,
  but less conveniently).
  
  The default degree is 2; degree 1.5 means inclusion of interactions but not 
  of quadratic terms. 
  
  The menu opens a response surface model selection menu, 
  providing a convenient starting formula.
}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
