\name{Menu.tab}
\alias{Menu.tab}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using the menu for tabulating selected factors}
\description{Help on using the menu for tabulating selected factors}
%- maybe also 'usage' for other objects documented here.
\section{Help on using the menu for tabulating selected factors}{
  This menu allows to select experimental factors for tabulating 
  (with command \code{\link{table}}). 
  
  The menu is particularly useful for viewing the structure of 
  general orthogonal arrays. 
  Up to three or four factors are most easily viewed together.
}

\seealso{See also \code{\link{table}} for the underlying function}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
