\name{Menus}
\alias{Menus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Help on using R-Commander menus}
\description{Important instructions for windows users for making the menus work well}
%- maybe also 'usage' for other objects documented here.
\section{Important Usage Note for Windows Users}{
The most important technical point about this menu and the R-Commander in general: \cr
The R GUI should be installed with the Single Document Interface (SDI) mode 
instead of the default MDI mode. John Fox, the author of R-Commander, has described 
how to change MDI to SDI, if R has been installed in the wrong mode 
(\url{http://socserv.socsci.mcmaster.ca/jfox/Misc/Rcmdr/installation-notes.html}). 
It is strongly discouraged to run the R-Commander under the MDI mode, 
as it happens very frequently that windows suddenly disappear. They usually 
have iconized only and can be retrieved from the taskbar; nevertheless, this behavior 
can be very annoying and can be avoided by using R in SDI mode.}

\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
