\name{RcmdrPlugin.DoEUndocumentedFunctions}
  \alias{RcmdrPlugin.DoEUndocumentedFunctions}
  \alias{activeDataSetDesign2P} 
  \alias{activeDataSetDesign2PResp}
  \alias{activeDataSetDesignlhs} 
  \alias{activeDataSetDesignLongToWide}
  \alias{activeDataSetDesignP} 
  \alias{activeDataSetDesignPResp} 
  \alias{activeDataSetDesignWide}
  \alias{dc} 
  \alias{display.design} 
  \alias{displayblock} 
  \alias{displayCentralCompositeDesigns}
  \alias{displayGlossary} 
  \alias{displaysplitplot} 
  \alias{exist2Designs} 
  \alias{existDesign2}
  \alias{existDesign2FrF} 
  \alias{existDesign2pb} 
  \alias{existDesigns} 
  \alias{existDesignsWithResp}
  \alias{help2level} 
  \alias{helpabout} 
  \alias{helpAnalyze} 
  \alias{helpExpImp} 
  \alias{helpgeneral} 
  \alias{helpGeneralAnalyze}
  \alias{onHelpTabEstimable}
  \alias{helpMenus}
  \alias{helpModify} 
  \alias{helpquant} 
  \alias{helpQuantDesignAnalyze} 
  \alias{HowToReImport}
  \alias{initializeDialogDoE} 
  \alias{isDesign2FrF} 
  \alias{isDesign2pb} 
  \alias{justDoItDoE} 
  \alias{listDesigns}
  \alias{listDesigns2} 
  \alias{listDesignsWithResp} 
  \alias{listLists} 
  \alias{listObjects} 
  \alias{Menu.aggregate}
  \alias{Menu.EffectPlots} 
  \alias{Menu.export} 
  \alias{Menu.responses}
  \alias{Menu.wide} 
  \alias{textcorrect} 
  \alias{textentry}
  
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Functions used in menus }
\description{Functions used in menus - these are not needed by users }
%- maybe also 'usage' for other objects documented here.
\usage{
  activeDataSetDesign2P() 
  activeDataSetDesign2PResp()
  activeDataSetDesignlhs() 
  activeDataSetDesignLongToWide()
  activeDataSetDesignP() 
  activeDataSetDesignPResp() 
  activeDataSetDesignWide()
  dc() 
  display.design() 
  displayblock() 
  displayCentralCompositeDesigns()
  displayGlossary() 
  displaysplitplot() 
  exist2Designs() 
  existDesign2()
  existDesign2FrF() 
  existDesign2pb() 
  existDesigns() 
  existDesignsWithResp()
  help2level() 
  helpabout()
  helpAnalyze() 
  helpExpImp() 
  helpgeneral() 
  helpGeneralAnalyze()
  onHelpTabEstimable()
  helpMenus()
  helpModify() 
  helpquant() 
  helpQuantDesignAnalyze() 
  HowToReImport()
  initializeDialogDoE(title = "", offset = 0, preventCrisp = FALSE) 
  isDesign2FrF(design) 
  isDesign2pb(design) 
  justDoItDoE(command) 
  listDesigns(envir = .GlobalEnv, ...)
  listDesigns2(envir = .GlobalEnv, type = NULL, ...) 
  listDesignsWithResp(envir = .GlobalEnv, ...) 
  listLists(envir = .GlobalEnv, ...) 
  listObjects(envir = .GlobalEnv, ...) 
  Menu.aggregate()
  Menu.EffectPlots() 
  Menu.export() 
  Menu.responses()
  Menu.wide() 
  textcorrect(message) 
  textentry()
}
\arguments{ 
\item{title}{dialog title}
\item{offset}{positioning parameter}
\item{preventCrisp}{ }
\item{design}{data frame of S3 class \code{\link[DoE.base:class-design]{design}}}
\item{message}{message to be displayed (character string)}
}


\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }
\keyword{ internal }% __ONLY ONE__ keyword per line
