\name{Menu.loadcatlg}
\alias{Menu.loadcatlg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Using the menu for loading catalogues}
\description{
  This help file describes loading a catalogue from a loaded catalogues package. 
  Currently, the only such package is FrF2.catlg128.
}
%- maybe also 'usage' for other objects documented here.
\details{
  This help file describes loading a catalogue from a loaded catalogues package. 
  Currently, the only such package is \pkg{FrF2.catlg128}.

  The menu is only available if package FrF2.catlg128 is loaded.
  
  Select the design from the drop-down list of available designs, and press OK.

  For importing an unloaded R work space first, use the File menu 
  (usually the left-most menu) of the R-commander. 
  
  For loading a catalogue package first (i.e. \pkg{FrF2.catlg128}), 
  use the menu entry for loading packages in the \emph{Extras} menu 
  of the R-commander. 
  
}
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
