\name{Menu.QuantDesignAnalyze}
\alias{Menu.QuantDesignAnalyze}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Analyze designs for quantitative factors}
\description{This is a brief explanation on the menu for analysis of designs for quantitative factors.}
%- maybe also 'usage' for other objects documented here.

\section{Available analysis options}{
Currently, no special analysis methods for quantitative designs have been implemented.
Use the analysis facilities within the R-commander, or use the command-line facilities 
offered e.g. in package \code{\link[rsm:rsm-package]{rsm}}
}


\references{ 
    Box G. E. P, Hunter, W. C. and Hunter, J. S. (2005) 
    \emph{Statistics for Experimenters, 2nd edition}.
    New York: Wiley.
 }
\author{ Ulrike Groemping }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ array }% __ONLY ONE__ keyword per line
